LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-delaynode-interface/idl-test-expected.txt

CONSOLE MESSAGE: line 482: callback not yet supported
CONSOLE MESSAGE: line 482: callback not yet supported
dictionary DelayOptions : AudioNodeOptions {
             double maxDelayTime = 1;
             double delayTime = 0;
};

[Constructor(BaseAudioContext context, optional DelayOptions options)]
interface DelayNode : AudioNode {

    readonly attribute AudioParam delayTime;

};

PASS webaudio Delay interfaces 
PASS DelayNode interface: existence and properties of interface object 
FAIL DelayNode interface object length assert_equals: wrong value for DelayNode.length expected 1 but got 0
PASS DelayNode interface object name 
PASS DelayNode interface: existence and properties of interface prototype object 
PASS DelayNode interface: existence and properties of interface prototype object's "constructor" property 
PASS DelayNode interface: existence and properties of interface prototype object's @@unscopables property 
PASS DelayNode interface: attribute delayTime 
PASS DelayNode must be primary interface of delay_node 
PASS Stringification of delay_node 
PASS DelayNode interface: delay_node must inherit property "delayTime" with the proper type 
